Fix scene crash: Add proper bounds checking + fix GateType padding
Two critical fixes:
1. Added comprehensive bounds checking before scene capture
2. Changed GateType from uint8_t to int32_t to eliminate struct padding

The combination of missing bounds checks and struct padding was causing crashes.

+++ src/deluge/model/clip/sequencer/modes/step_sequencer_mode.h
@@ -85,7 +85,7 @@ private:
 	static constexpr int32_t kMaxScaleNotes = 32;
 
 	// Gate types for each step
-	enum class GateType : uint8_t {
+	enum class GateType : int32_t {
 		OFF = 0,   // No note, but step duration counts
 		ON = 1,    // Play note
 		SKIP = 2,  // Skip step entirely, jump to next
